Update BVM-fullworkingcode.ino
Changed the display from an Adafruit_SSD1351 to an Adafruit_STT7735. There are still graphical errors on the screen, however the numbers update as the flow sensor reads the data. Graphical errors included pixelated background with multiple colors, the numbers being printed for the flow sensor taking up more than one line of the display, and the word danger, which should be used only when the number reaches over 600, is constantly on display at the bottom of the screen.