Current Code
These are some of the issues with the code. The alarm sounds as soon as you switch on the device ( I commented that part of the code out at the end). I want the bars to be filled with different colors. The bottom bar with blue ( insufficient volume) the middle bar with green( just right) the top bar with red ( too much). I have a code that fills the background with those colors and I commented out. The other code is filling the bars with two colors and missing the third color. I also commented it out. Also, I am trying to have a line that goes up to be the guide to the user.